[BE-85] refactor: imagefile 패키지 이름을 file로 변경(후에 다른 종류의 파일을 받을 수 있어서 정확히 명시함)
