+++ src/main/java/com/recordit/server/exception/file/FileInputStreamException.java
@@ -1,4 +1,4 @@
-package com.recordit.server.exception.imagefile;
+package com.recordit.server.exception.file;
 
 public class FileInputStreamException extends RuntimeException {
 	public FileInputStreamException(String message) {
